database setup sql script removes long column, adds short column

+++ create_table.sql
@@ -2,7 +2,7 @@ DROP TABLE IF EXISTS urls;
 
 CREATE TABLE urls (
   id INTEGER PRIMARY KEY AUTOINCREMENT,
-  long TEXT NOT NULL,
+  short TEXT NOT NULL,
   desktop TEXT,
   mobile TEXT,
   tablet TEXT,
@@ -10,5 +10,5 @@ CREATE TABLE urls (
   updatedAt DATETIME NOT NULL
 );
 
-INSERT INTO urls (long, desktop, mobile, tablet, createdAt, updatedAt)
+INSERT INTO urls (short, desktop, mobile, tablet, createdAt, updatedAt)
 VALUES ('a', 'b', 'c', 'd', datetime(), datetime());
